vérifs à terminer

+++ src/Controller/UserProfilController.php
@@ -49,40 +49,44 @@ class UserProfilController extends AbstractController
     
             if($form->isSubmitted() && $form->isValid())
             {
-                if($form->get('plainPassword')->getData() !== null) 
+                if(!is_null($form['pseudo']->getdata()) && !is_null($form['plainPassword']->getdata()) && !is_null($form['metier']->getdata()) && !is_null($form['niveau']->getdata()))
                 {
-                    $user->setPassword(
-                        $passwordEncoder->encodePassword(
-                            $user,
-                            $form->get('plainPassword')->getData()
-                        )
-                    );
-                }
-    
-                // image profil
-    
-                
-                $infoImg1 = $form['imgprofil']->getData();
-                $oldNomImg1 = $user->getImgprofil();
-    
-                if($infoImg1!=null){
-                    $oldCheminImg1 = $this->getParameter('photos_users') . '/' . $oldNomImg1;       
-                    if (file_exists($oldCheminImg1)) 
+                    if($form->get('plainPassword')->getData() !== null) 
                     {
-                        unlink($oldCheminImg1);
+                        $user->setPassword(
+                            $passwordEncoder->encodePassword(
+                                $user,
+                                $form->get('plainPassword')->getData()
+                            )
+                        );
                     }
-                    $extensionImg1 = $infoImg1->guessExtension();
-                    $nomImg1 = '1-' . time() . '.' . $extensionImg1;
-                    $infoImg1->move($this->getParameter('photos_users'), $nomImg1);
-                    $user->setImgprofil($nomImg1);
+        
+                    // image profil
+        
+                    
+                    $infoImg1 = $form['imgprofil']->getData();
+                    $oldNomImg1 = $user->getImgprofil();
+        
+                    if($infoImg1!=null){
+                        $oldCheminImg1 = $this->getParameter('photos_users') . '/' . $oldNomImg1;       
+                        if (file_exists($oldCheminImg1)) 
+                        {
+                            unlink($oldCheminImg1);
+                        }
+                        $extensionImg1 = $infoImg1->guessExtension();
+                        $nomImg1 = '1-' . time() . '.' . $extensionImg1;
+                        $infoImg1->move($this->getParameter('photos_users'), $nomImg1);
+                        $user->setImgprofil($nomImg1);
+                    }
+                    
+        
+                    $manager = $this->getDoctrine()->getManager();
+                    $manager->persist($user);
+                    $manager->flush();
+                    $this->addFlash('success', 'Vous avez modifié votre profil!');
+                    // redirection vers profil user courant 
+                    return $this->redirectToRoute('user_profil', ['id' => $id]);
                 }
-                
-    
-                $manager = $this->getDoctrine()->getManager();
-                $manager->persist($user);
-                $manager->flush();
-                // redirection vers profil user courant 
-                return $this->redirectToRoute('user_profil', ['id' => $id]);
             }
             return $this->render('user_profil/userProfilForm.html.twig', [
                 'userProfilForm' => $form->createView()
